Finish session 4.

+++ bft_interp/src/lib.rs
@@ -10,8 +10,13 @@ use bft_types::BFprogram;
 #[allow(dead_code)]
 #[derive(Debug)]
 pub struct BFVM<C> {
+    /// Block of memory for the program to work on.
     tape: Vec<C>,
+
+    /// Index of where the program is pointing to in the tape.
     head: usize,
+
+    /// When true, the VM is allowed to grow the tape for additional space as needed.
     growable: bool,
 }
 
@@ -23,21 +28,25 @@ impl<C: Default> BFVM<C> {
     pub fn new(capacity: Option<NonZeroUsize>, growable: bool) -> BFVM<C> {
         let c = capacity.map_or(30000, NonZeroUsize::get);
         let mut tape = Vec::new();
-        tape.resize_with(usize::from(c), C::default);
+        tape.resize_with(c, C::default);
         BFVM {
             tape,
             head: 0,
             growable,
         }
     }
-
 }
 
 impl<C> BFVM<C> {
     /// The main interpreter that takes a program and (eventually) interprets it.
     pub fn interpret(&self, code: &BFprogram) {
         for inst in code.instructions() {
-            println!("[{:?}:{}] {}", code.source(), inst.location(), inst.instruction());
+            println!(
+                "[{:?}:{}] {}",
+                code.source(),
+                inst.location(),
+                inst.instruction()
+            );
         }
     }
 }
